refactor: 나의 활동 화면 UI별 분리 (#19)
